Impedance mismatch

+++ day1/01-starting/Bookings.Domain/Money.cs
@@ -1,14 +1,14 @@
 namespace Bookings.Domain;
 
 public record Money {
-    public float  Amount   { get; internal init; }
+    public decimal  Amount   { get; internal init; }
     public string Currency { get; internal init; }
 
     static readonly string[] SupportedCurrencies = {"USD", "EUR", "GPB"};
 
     internal Money() { }
 
-    public Money(float amount, string currency) {
+    public Money(decimal amount, string currency) {
         if (!SupportedCurrencies.Contains(currency)) throw new DomainException($"Unsupported currency: {currency}");
 
         Amount   = amount;
@@ -23,5 +23,5 @@ public record Money {
         return new Money(one.Amount - another.Amount, one.Currency);
     }
 
-    public static implicit operator double(Money money) => money.Amount;
+    public static implicit operator decimal(Money money) => money.Amount;
 }
